PyLabel/PyLabelProgramFiles/requirements.txt
Pillow==9.5.0
brother_ql